apps/Debt-Collection/requirements.txt
Flask>=3.0
python-dotenv>=1.0
google-api-python-client>=2.143.0
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
